requirements.txt
cocos2d==0.6.5
Cython==0.28.2
future==0.17.1
glfw==1.7.1
imgui==0.1.0
numpy==1.16.1
pgzero==1.2
pygame==1.9.4
pyglet==1.3.2
PyOpenGL==3.1.0
PySDL2==0.9.6
six==1.12.0
